+++ tests/sandbox_tests/test_wait_helper.py
@@ -11,8 +11,8 @@ class WaitHelpersTestCase(SandboxedNodeTestCase):
     def test_1_send_multiple_transactions(self) -> None:
         client = self.client
         operations = [
-            client.transaction(destination=sandbox_addresses['bootstrap3'], amount=1000).send()
-            for _ in range(2)
+            self.get_client(key).transaction(destination=sandbox_addresses['bootstrap5'], amount=1000).send()
+            for key in ['bootstrap1', 'bootstrap2']
         ]
         with self.assertRaises(TimeoutError):
             client.wait(*operations, num_blocks_wait=1, time_between_blocks=1, block_timeout=2)
@@ -20,8 +20,8 @@ class WaitHelpersTestCase(SandboxedNodeTestCase):
         self.bake_block()
 
         operations.extend([
-            client.transaction(destination=sandbox_addresses['bootstrap3'], amount=1000).send()
-            for _ in range(2)
+            self.get_client(key).transaction(destination=sandbox_addresses['bootstrap5'], amount=1000).send()
+            for key in ['bootstrap3', 'bootstrap4']
         ])
         with self.assertRaises(TimeoutError):
             client.wait(*operations, num_blocks_wait=1, time_between_blocks=1, block_timeout=2)
@@ -30,5 +30,5 @@ class WaitHelpersTestCase(SandboxedNodeTestCase):
         self.bake_block()
 
         client.wait(*operations, num_blocks_wait=1, time_between_blocks=1, prev_hash=head_hash)
-        bootstrap3 = self.client.shell.contracts[sandbox_addresses['bootstrap3']]()
-        self.assertEqual('4000000004000', bootstrap3['balance'])
+        bootstrap5 = self.client.shell.contracts[sandbox_addresses['bootstrap5']]()
+        self.assertEqual('3800000004000', bootstrap5['balance'])