+++ tests/unit_tests/test_michelson/test_repl/test_opcodes.py
@@ -664,6 +664,18 @@ class OpcodesTestCase(TestCase):
             '(Pair 2310000 1010000)',
             '(Some (Pair 3320000 1300000))',
         ),
+        (
+            'tez_add_sub.tz',
+            'None',
+            '(Pair 2000000 1000000)',
+            '(Some (Pair 3000000 1000000))',
+        ),
+        (
+            'tez_add_sub.tz',
+            'None',
+            '(Pair 2310000 1010000)',
+            '(Some (Pair 3320000 1300000))',
+        ),
         # Test various additions
         ('add.tz', 'Unit', 'Unit', 'Unit'),
         # Test ABS