+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 3.4.2 - 2022-03-26
+
+### Fixed
+
+* time_between_blocks no longer exists, minimum_block_delay used instead
+* sandbox tests -> re-create node container per each case (cannot rollback in Tenderbake)
+* REPL: `CREATE_CONTRACT` is now able to originate contracts with views (by @konchunas)
+
+### Added
+
+* `SUB_MUTEZ` instruction support in REPL
+
 ## 3.4.1 - 2022-03-22
 
 ### Fixed